api/requirements.txt
fastapi==0.115.12
duckdb==1.2.1
pydantic==2.11.2
mangum==0.19.0
supabase==2.3.4
google-generativeai
upstash-redis==1.2.0
python-dotenv==1.0.1
